Add special handling of 0 for READ/WRITE (6) to SCSI-to-USB bridge

+++ cpp/s2pdump/s2pdump_core.h
@@ -32,7 +32,7 @@ public:
     using ScsiDeviceInfo = struct {
         bool removable;
         byte type;
-        byte scsi_level;
+        int scsi_level;
         string vendor;
         string product;
         string revision;
